pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fund-screener"
version = "0.1.0"
description = "Production-grade financial analysis pipeline"
authors = [
    {name = "Fund Screener Team", email = "team@fundscreener.com"},
]
dependencies = [
    "pandas>=2.0.0",
    "yfinance>=0.2.0",
    "pydantic>=2.0.0",
    "typer[all]>=0.9.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF"]
ignore = ["S101", "PLR0913", "PLR0912", "PLR0915"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
